Restructurize ProductListView
