+++ Test/Test5/Models/ProductModel.swift
@@ -1,0 +1,26 @@
+//
+//  ProductListModel.swift
+//  Test
+//
+//  Created by Aditya Cahyo on 25/06/23.
+//
+
+import Foundation
+
+struct ProductResponse: Decodable {
+    var responseCode: Int
+    var responseDesc: String
+    var data: [Product]
+    
+    struct Product: Decodable, Identifiable {
+        var productId: Int
+        var productName: String
+        var productImage: String
+        var productDesc: String
+        var productPrice: Int
+        
+        var id: Int {
+            return productId
+        }
+    }
+}
